Fix instrument names
