+++ src/qibolab/tests/test_instruments_qutech.py
@@ -3,43 +3,26 @@ import pytest
 from qibolab import Platform
 from qibolab.instruments.qutech import SPI
 from qibolab.paths import user_folder
+from qibolab.tests.utils import InstrumentsDict, load_from_platform
 
 INSTRUMENTS_LIST = ["SPI"]
 
 
-class InstrumentsDict(dict):
-    def __getitem__(self, name):
-        if name not in self:
-            pytest.skip(f"Skip {name} test as it is not included in the tested platforms.")
-        else:
-            return super().__getitem__(name)
-
-
 instruments = InstrumentsDict()
+instruments_settings = {}
 
 # To test --> name = SpiRack
 @pytest.mark.qpu
 @pytest.mark.parametrize("name", INSTRUMENTS_LIST)
 def test_instruments_qutech_init(platform_name, name):
     settings = Platform(platform_name).settings
-
-    if name not in settings["instruments"]:
-        pytest.skip(f"Skip {name} test as it is not included in the tested platforms.")
-
     # Instantiate instrument
-    lib = settings["instruments"][name]["lib"]
-    i_class = settings["instruments"][name]["class"]
-    address = settings["instruments"][name]["address"]
-    from importlib import import_module
-
-    InstrumentClass = getattr(import_module(f"qibolab.instruments.{lib}"), i_class)
-    instance = InstrumentClass(name, address)
+    instance, instr_settings = load_from_platform(settings, name)
     instruments[name] = instance
+    instruments_settings[name] = instr_settings
     assert instance.name == name
-    assert instance.address == address
     assert instance.is_connected == False
     assert instance.device == None
-    assert instance.signature == f"{name}@{address}"
     assert instance.data_folder == user_folder / "instruments" / "data" / instance.tmp_folder.name.split("/")[-1]
 
 
@@ -53,7 +36,7 @@ def test_instruments_qutech_connect(name):
 @pytest.mark.parametrize("name", INSTRUMENTS_LIST)
 def test_instruments_qutech_setup(platform_name, name):
     settings = Platform(platform_name).settings
-    instruments[name].setup(**settings["settings"], **settings["instruments"][name]["settings"])
+    instruments[name].setup(**settings["settings"], **instruments_settings[name])
 
 
 @pytest.mark.qpu
