+++ src/qibolab/tests/test_instruments_rohde_schwarz.py
@@ -4,43 +4,25 @@ import pytest
 from qibolab import Platform
 from qibolab.instruments.rohde_schwarz import SGS100A
 from qibolab.paths import user_folder
+from qibolab.tests.utils import InstrumentsDict, load_from_platform
 
 INSTRUMENTS_LIST = ["SGS100A"]
 
-
-class InstrumentsDict(dict):
-    def __getitem__(self, name):
-        if name not in self:
-            pytest.skip(f"Skip {name} test as it is not included in the tested platforms.")
-        else:
-            return super().__getitem__(name)
-
-
 instruments = InstrumentsDict()
+instruments_settings = {}
 
 
 @pytest.mark.qpu
 @pytest.mark.parametrize("name", INSTRUMENTS_LIST)
 def test_instruments_rohde_schwarz_init(platform_name, name):
     settings = Platform(platform_name).settings
-
-    if name not in settings["instruments"]:
-        pytest.skip(f"Skip {name} test as it is not included in the tested platforms.")
-
     # Instantiate instrument
-    lib = settings["instruments"][name]["lib"]
-    i_class = settings["instruments"][name]["class"]
-    address = settings["instruments"][name]["address"]
-    from importlib import import_module
-
-    InstrumentClass = getattr(import_module(f"qibolab.instruments.{lib}"), i_class)
-    instance = InstrumentClass(name, address)
+    instance, instr_settings = load_from_platform(settings, name)
     instruments[name] = instance
+    instruments_settings[name] = instr_settings
     assert instance.name == name
-    assert instance.address == address
     assert instance.is_connected == False
     assert instance.device == None
-    assert instance.signature == f"{name}@{address}"
     assert instance.data_folder == user_folder / "instruments" / "data" / instance.tmp_folder.name.split("/")[-1]
 
 
@@ -54,9 +36,9 @@ def test_instruments_rohde_schwarz_connect(name):
 @pytest.mark.parametrize("name", INSTRUMENTS_LIST)
 def test_instruments_rohde_schwarz_setup(platform_name, name):
     settings = Platform(platform_name).settings
-    instruments[name].setup(**settings["settings"], **settings["instruments"][name]["settings"])
-    for parameter in settings["instruments"][name]["settings"]:
-        assert getattr(instruments[name], parameter) == settings["instruments"][name]["settings"][parameter]
+    instruments[name].setup(**settings["settings"], **instruments_settings[name])
+    for parameter in instruments_settings[name]:
+        assert getattr(instruments[name], parameter) == instruments_settings[name][parameter]
 
 
 def instrument_set_and_test_parameter_values(instrument, parameter, values):
